new pause menu

+++ c/bt/controllers.h
@@ -26,8 +26,9 @@ typedef struct {
   bt_controller_buttons_t pressed;
   bt_controller_buttons_t released;
   struct {
-    f32 x;
-    f32 y;
+    // should be floats, but floats are broken at the moment
+    u32 x;
+    u32 y;
   } joystick;
 } bt_controller_t;
 #define bt_controllers ((bt_controller_t*)0x80079AF8)
